requirements.txt
django==4.1.4
pytest-django==4.5.2
